Extracts AndroidViewEnvironment
