+++ workflow-ui/container-android/src/main/java/com/squareup/workflow1/ui/backstack/BackStackContainer.kt
@@ -20,7 +20,6 @@ import com.squareup.workflow1.ui.BuilderViewFactory
 import com.squareup.workflow1.ui.Named
 import com.squareup.workflow1.ui.ViewEnvironment
 import com.squareup.workflow1.ui.ViewFactory
-import com.squareup.workflow1.ui.ViewRegistry
 import com.squareup.workflow1.ui.WorkflowUiExperimentalApi
 import com.squareup.workflow1.ui.androidx.WorkflowAndroidXSupport.stateRegistryOwnerFromViewTreeOrContext
 import com.squareup.workflow1.ui.androidx.WorkflowLifecycleOwner
@@ -98,9 +97,8 @@ public open class BackStackContainer @JvmOverloads constructor(
         return
       }
 
-    val newView = environment[ViewRegistry].buildView(
+    val newView = environment.buildView(
       initialRendering = named.top,
-      initialViewEnvironment = environment,
       contextForNewView = this.context,
       container = this,
       initializeView = {
